+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [12.1.0] - 2024-07-11
+
+### Highlights
+- The following Node methods now have a new recurse parameter that controls whether or not the method should be applied to descendant nodes:
+  - scale
+  - recolor
+  - remapResources
+- The following Node generator methods now have a new recurse parameter that controls whether or not the method should yield objects of descendant nodes:
+  - walkEffects
+  - walkActions
+  - walkProperties
+- This new recurse parameter for the Node methods defaults to true, so calling them without it works just like before this change.
+
 ## [12.0.0] - 2024-06-27
 
 ### Highlights
@@ -52,5 +65,6 @@
 - External values 2000 and 70200 for AC6 have been documented thanks to lugia19.
 - Fixed action 301 (EqualDistanceEmitter) missing a type for one of its fields, potentially causing issues when writing to DS3's structure.
 
+[12.1.0]: https://github.com/EvenTorset/fxr/compare/v12.0.0...v12.1.0
 [12.0.0]: https://github.com/EvenTorset/fxr/compare/v11.0.0...v12.0.0
 [11.0.0]: https://github.com/EvenTorset/fxr/compare/v10.0.1...v11.0.0